+++ velox/functions/sparksql/aggregates/CMakeLists.txt
@@ -26,7 +26,8 @@ velox_add_library(
   Register.cpp
   RegrReplacementAggregate.cpp
   SumAggregate.cpp
-  VarianceAggregate.cpp)
+  VarianceAggregate.cpp
+  CovarianceAggregate.cpp)
 
 velox_link_libraries(
   velox_functions_spark_aggregates
